Bug-fix

+++ jiLan-his/src/main/java/com/acmed/his/model/dto/WaterDayAndMonthCountDto.java
@@ -3,6 +3,8 @@ package com.acmed.his.model.dto;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * WaterDayAndMonthCount
  * 昨日收益  本月收益
@@ -14,7 +16,7 @@ public class WaterDayAndMonthCountDto {
     @ApiModelProperty("机构编号")
     private Integer orgCode;
     @ApiModelProperty("昨日收益")
-    private Integer dayTotal;
+    private BigDecimal dayTotal;
     @ApiModelProperty("本月收益")
-    private Integer monthTotal;
+    private BigDecimal monthTotal;
 }
